+++ src/MQTTPersistence.h
@@ -39,10 +39,10 @@ int MQTTPersistence_initialize(Clients* c, const char* serverURI);
 int MQTTPersistence_close(Clients* c);
 int MQTTPersistence_clear(Clients* c);
 int MQTTPersistence_restore(Clients* c);
-void* MQTTPersistence_restorePacket(char* buffer, int buflen);
-void MQTTPersistence_insertInOrder(List* list, void* content, int size);
-int MQTTPersistence_put(int socket, char* buf0, int buf0len, int count, 
-								 char** buffers, int* buflens, int htype, uint64_t msgId, int scr);
+void* MQTTPersistence_restorePacket(char* buffer, size_t buflen);
+void MQTTPersistence_insertInOrder(List* list, void* content, size_t size);
+int MQTTPersistence_put(int socket, char* buf0, size_t buf0len, int count, 
+								 char** buffers, size_t* buflens, int htype, uint64_t msgId, int scr);
 int MQTTPersistence_remove(Clients* c, char* type, int qos, uint64_t msgId);
 void MQTTPersistence_wrapMsgID(Clients *c);
